Rename Smaily namespace to SmailyForOpenCart
